refactor(step5): step 패키지명을 구체적으로 변경
