+++ src/main/kotlin/racingcar/domain/Mover.kt
@@ -1,0 +1,6 @@
+package racingcar.domain
+
+interface Mover {
+
+    fun canMove(): Boolean
+}
