+++ src/main/kotlin/racingcar/domain/RacingRecord.kt
@@ -1,0 +1,6 @@
+package racingcar.domain
+
+data class RacingRecord(
+    val carName: String,
+    val carDistance: Int
+)
