+++ src/main/kotlin/racingcar/domain/RacingResult.kt
@@ -1,0 +1,5 @@
+package racingcar.domain
+
+data class RacingResult(
+    val records: List<RacingRecord>
+)
