--- src/main/kotlin/step2/view/InputView.kt
@@ -1,4 +1,0 @@
-package step2.view
-
-fun getSplitStringForCalculate() = readLine()?.split(" ")
-    ?: throw IllegalArgumentException("읽은 문자열은 null 일 수 없음.")
