--- src/main/kotlin/step2/view/OutputView.kt
@@ -1,3 +1,0 @@
-package step2.view
-
-fun printInsertStringForCalculate() = println("사칙연산 문자열을 입력해주세요 (띄어쓰기로 구분)")
