--- src/main/kotlin/step3/domain/CarMover.kt
@@ -1,16 +1,0 @@
-package step3.domain
-
-import kotlin.random.Random
-import kotlin.random.nextInt
-
-class CarMover : Mover {
-
-    override fun canMove(): Boolean {
-        return Random.nextInt(0..MAX_RANDOM_NUM) >= CONDITION_CAN_MOVE
-    }
-
-    companion object {
-        private const val MAX_RANDOM_NUM = 9
-        private const val CONDITION_CAN_MOVE = 4
-    }
-}
