--- src/main/kotlin/step3/domain/Cars.kt
@@ -1,22 +1,0 @@
-package step3.domain
-
-class Cars(
-    private val cars: List<Car>
-) {
-    fun race() {
-        cars.forEach { it.race() }
-    }
-
-    fun getCurrentRacingRecords(): List<RacingRecord> {
-        return cars.map { RacingRecord(it.name, it.distance) }
-    }
-
-    fun findCurrentWinners(): List<String> {
-        val furthestDistance = getFurthestDistance()
-        return cars
-            .filter { it.distance == furthestDistance }
-            .map { it.name }
-    }
-
-    private fun getFurthestDistance() = cars.map { it.distance }.max()
-}
