--- src/main/kotlin/step3/domain/Mover.kt
@@ -1,6 +1,0 @@
-package step3.domain
-
-interface Mover {
-
-    fun canMove(): Boolean
-}
