--- src/main/kotlin/step3/domain/RacingRecord.kt
@@ -1,6 +1,0 @@
-package step3.domain
-
-data class RacingRecord(
-    val carName: String,
-    val carDistance: Int
-)
