--- src/main/kotlin/step3/domain/RacingResult.kt
@@ -1,5 +1,0 @@
-package step3.domain
-
-data class RacingResult(
-    val records: List<RacingRecord>
-)
