+++ src/main/kotlin/stringcalculator/StringCalculatorApplication.kt
@@ -1,0 +1,10 @@
+package stringcalculator
+
+import stringcalculator.view.getSplitStringForCalculate
+import stringcalculator.view.printInsertStringForCalculate
+
+fun main() {
+    printInsertStringForCalculate()
+    val symbols = getSplitStringForCalculate()
+    println(StringCalculator.calculate(symbols))
+}
