+++ src/main/kotlin/stringcalculator/view/OutputView.kt
@@ -1,0 +1,3 @@
+package stringcalculator.view
+
+fun printInsertStringForCalculate() = println("사칙연산 문자열을 입력해주세요 (띄어쓰기로 구분)")
